MOO-221: moved dhis stuffs from MSFCoreService into a DHISService
