+++ api/src/main/java/org/openmrs/module/msfcore/MSFCoreActivator.java
@@ -28,6 +28,7 @@ import org.openmrs.module.idgen.service.IdentifierSourceService;
 import org.openmrs.module.metadatadeploy.api.MetadataDeployService;
 import org.openmrs.module.metadatamapping.MetadataTermMapping;
 import org.openmrs.module.metadatamapping.api.MetadataMappingService;
+import org.openmrs.module.msfcore.api.DHISService;
 import org.openmrs.module.msfcore.api.MSFCoreService;
 import org.openmrs.module.msfcore.metadata.MSFMetadataBundle;
 import org.openmrs.module.msfcore.metadata.PatientIdentifierTypes;
@@ -48,7 +49,7 @@ public class MSFCoreActivator extends BaseModuleActivator {
     public void started() {
         log.info("Started MSF Core Module");
 
-        Context.getService(MSFCoreService.class).transferDHISMappingsToDataDirectory();
+        Context.getService(DHISService.class).transferDHISMappingsToDataDirectory();
 
         installMSFMeta();
