+++ api/src/main/java/org/openmrs/module/msfcore/api/DHISService.java
@@ -1,0 +1,32 @@
+/**
+ * This Source Code Form is subject to the terms of the Mozilla Public License,
+ * v. 2.0. If a copy of the MPL was not distributed with this file, You can
+ * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
+ * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
+ * 
+ * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
+ * graphic logo is a trademark of OpenMRS Inc.
+ */
+package org.openmrs.module.msfcore.api;
+
+import java.util.Properties;
+
+import org.openmrs.Location;
+import org.openmrs.LocationAttribute;
+import org.openmrs.Patient;
+import org.openmrs.api.OpenmrsService;
+
+/**
+ * This is a DSHI2 synchronisation service
+ */
+public interface DHISService extends OpenmrsService {
+    public String getLocationDHISUid(Location location);
+
+    public LocationAttribute getLocationUidAttribute(Location location);
+
+    public void transferDHISMappingsToDataDirectory();
+
+    public Properties getDHISMappings();
+
+    public String postTrackerInstanceThroughOpenHimForAPatient(Patient patient);
+}
