+++ api/src/main/java/org/openmrs/module/msfcore/api/MSFCoreService.java
@@ -10,13 +10,10 @@
 package org.openmrs.module.msfcore.api;
 
 import java.util.List;
-import java.util.Properties;
 
 import org.openmrs.Concept;
 import org.openmrs.Location;
 import org.openmrs.LocationAttribute;
-import org.openmrs.LocationAttributeType;
-import org.openmrs.Patient;
 import org.openmrs.api.OpenmrsService;
 import org.openmrs.module.idgen.SequentialIdentifierGenerator;
 import org.openmrs.module.msfcore.DropDownFieldOption;
@@ -28,8 +25,6 @@ import org.openmrs.module.msfcore.DropDownFieldOption;
 public interface MSFCoreService extends OpenmrsService {
 
     public List<Concept> getAllConceptAnswers(Concept question);
-
-    public List<LocationAttribute> getLocationAttributeByTypeAndLocation(LocationAttributeType type, Location location);
 
     public List<DropDownFieldOption> getAllConceptAnswerNames(String uuid);
 
@@ -51,13 +46,4 @@ public interface MSFCoreService extends OpenmrsService {
 
     public void saveSequencyPrefix(SequentialIdentifierGenerator generator);
 
-    public String getLocationDHISUid(Location location);
-
-    public LocationAttribute getLocationUidAttribute(Location location);
-
-    public void transferDHISMappingsToDataDirectory();
-
-    public Properties getDHISMappings();
-
-    public String postTrackerInstanceThroughOpenHimForAPatient(Patient patient);
 }
