+++ api/src/main/resources/moduleApplicationContext.xml
@@ -59,6 +59,23 @@
             <ref bean="transactionAttributeSource" />
         </property>
     </bean>
+    <bean id="dhisService"
+            class="org.springframework.transaction.interceptor.TransactionProxyFactoryBean">
+        <property name="transactionManager">
+            <ref bean="transactionManager" />
+        </property>
+        <property name="target">
+            <bean class="org.openmrs.module.msfcore.api.impl.DHISServiceImpl">
+                <property name="dao" ref="msfcore.MSFCoreDao" />
+            </bean>
+        </property>
+        <property name="preInterceptors">
+            <ref bean="serviceInterceptors" />
+        </property>
+        <property name="transactionAttributeSource">
+            <ref bean="transactionAttributeSource" />
+        </property>
+    </bean>
 
     <bean parent="serviceContext">
         <property name="moduleService">
@@ -68,12 +85,19 @@
             </list>
         </property>
     </bean>
-    
     <bean parent="serviceContext">
         <property name="moduleService">
             <list>
                 <value>org.openmrs.module.msfcore.api.AuditService</value>
                 <ref bean="auditService" />
+            </list>
+        </property>
+    </bean>
+    <bean parent="serviceContext">
+        <property name="moduleService">
+            <list>
+                <value>org.openmrs.module.msfcore.api.DHISService</value>
+                <ref bean="dhisService" />
             </list>
         </property>
     </bean>
