+++ omod/src/main/java/org/openmrs/module/msfcore/page/controller/ConfigurationPageController.java
@@ -12,6 +12,7 @@ import org.openmrs.Location;
 import org.openmrs.LocationAttribute;
 import org.openmrs.api.LocationService;
 import org.openmrs.module.msfcore.SimplifiedLocation;
+import org.openmrs.module.msfcore.api.DHISService;
 import org.openmrs.module.msfcore.api.MSFCoreService;
 import org.openmrs.ui.framework.UiUtils;
 import org.openmrs.ui.framework.annotation.SpringBean;
@@ -23,8 +24,8 @@ public class ConfigurationPageController {
     public void controller(PageModel model, @RequestParam(value = "defaultLocationUuid", required = false) Location defaultLocation,
                     @RequestParam(value = "instanceId", required = false) String instanceId,
                     @SpringBean("locationService") LocationService locationService,
-                    @SpringBean("msfCoreService") MSFCoreService msfCoreService, HttpServletRequest request, HttpServletResponse response,
-                    UiUtils ui) throws IOException {
+                    @SpringBean("msfCoreService") MSFCoreService msfCoreService, @SpringBean("dhisService") DHISService dhisService,
+                    HttpServletRequest request, HttpServletResponse response, UiUtils ui) throws IOException {
         boolean isPostRequest = false;
         if (defaultLocation != null) {
             isPostRequest = true;
@@ -34,10 +35,10 @@ public class ConfigurationPageController {
             isPostRequest = true;
             msfCoreService.saveInstanceId(StringUtils.deleteWhitespace(instanceId).toUpperCase());
         }
-        saveLocationCodes(request, msfCoreService, locationService, isPostRequest);
+        saveLocationCodes(request, msfCoreService, dhisService, locationService, isPostRequest);
         model.addAttribute("defaultLocation", locationService.getDefaultLocation());
         model.addAttribute("allLocations", locationService.getAllLocations());
-        model.addAttribute("msfLocations", getSimplifiedMSFLocations(msfCoreService));
+        model.addAttribute("msfLocations", getSimplifiedMSFLocations(msfCoreService, dhisService));
         model.addAttribute("instanceId", msfCoreService.instanceId());
         if (isPostRequest && msfCoreService.configured()) {
             // reload msfIDgenerator installation
@@ -46,20 +47,20 @@ public class ConfigurationPageController {
         }
     }
 
-    private List<SimplifiedLocation> getSimplifiedMSFLocations(MSFCoreService msfCoreService) {
+    private List<SimplifiedLocation> getSimplifiedMSFLocations(MSFCoreService msfCoreService, DHISService dhisService) {
         List<SimplifiedLocation> locations = new ArrayList<SimplifiedLocation>();
         for (Location loc : msfCoreService.getMSFLocations()) {
-            locations.add(new SimplifiedLocation(loc.getName(), msfCoreService.getLocationCode(loc), loc.getUuid(), msfCoreService
+            locations.add(new SimplifiedLocation(loc.getName(), msfCoreService.getLocationCode(loc), loc.getUuid(), dhisService
                             .getLocationDHISUid(loc)));
         }
         return locations;
     }
 
-    private void saveLocationCodes(HttpServletRequest request, MSFCoreService msfCoreService, LocationService locationService,
-                    boolean isPostRequest) {
+    private void saveLocationCodes(HttpServletRequest request, MSFCoreService msfCoreService, DHISService dhisService,
+                    LocationService locationService, boolean isPostRequest) {
         for (Location loc : msfCoreService.getMSFLocations()) {
             SimplifiedLocation simplifiedLocation = new SimplifiedLocation(loc.getName(), msfCoreService.getLocationCode(loc), loc
-                            .getUuid(), msfCoreService.getLocationDHISUid(loc));
+                            .getUuid(), dhisService.getLocationDHISUid(loc));
             Location location = locationService.getLocationByUuid(simplifiedLocation.getUuid());
             String code = request.getParameter(simplifiedLocation.getUuid());
             String uid = request.getParameter(simplifiedLocation.getUuid() + "_uid");
@@ -69,7 +70,7 @@ public class ConfigurationPageController {
                                 locationAttribute, location);
             }
             if (StringUtils.isNotBlank(uid) && !StringUtils.deleteWhitespace(uid).equals(simplifiedLocation.getUid())) {
-                LocationAttribute locationAttribute = msfCoreService.getLocationUidAttribute(location);
+                LocationAttribute locationAttribute = dhisService.getLocationUidAttribute(location);
                 saveLocationAttribute(isPostRequest, msfCoreService, locationService, simplifiedLocation, uid, locationAttribute, location);
             }
         }
